Added psydac/core/arrays.py from stefan-psydac
